+++ starkcore/__init__.py
@@ -4,7 +4,7 @@ if sys.version_info.major == 2:
     reload(sys)
     sys.setdefaultencoding("UTF8")
 
-version = "0.3.2"
+version = "0.4.0"
 
 from .environment import Environment
 from .user import Organization, Project, User