+++ CHANGELOG.md
@@ -13,6 +13,10 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Changed
+- raw methods response type
+### Fixed
+- prefix param
 
 ## [0.3.2] - 2024-05-27
 ### Fixed